// src/app/interfaces/user.ts
export interface User {
    firstName: string;
    lastName: string;
    email: string;
    mobile: string;
    subject: string;
    message: string;
    _id: string;
    submittedAt: Date;
}

// export interface DoNotSellMyInfoUser {
//     firstName: string;
//     lastName: string;
//     company: string;
//     email: string;
//     jobTitle: string;
//     country: string;
//     _id: string;
// }